refactor: Refactor the access to the countryCodes DB into its own module
To clean up the main server.js file the access to the files with the country
names and the country codes has been moved to its own file country-info-service.js
